+++ shipsay/configs/config.ini.php
@@ -27,10 +27,12 @@ $use_gzip = 1;
 $enable_down = 0;
 $is_ft = 0; 
 
-$theme_dir = '2025';
+$theme_dir = 'shipsay';
 $is_3in1 = 0;
 $commend_ids = '1113, 222, 3333, 4444, 5555, 6666, 7222, 8333, 9444, 10555';
 $category_per_page = 20;
+$home_lastupdate_num = 30;
+$home_postdate_num = 30;
 $readpage_split_mode = 2;
 $readpage_split_lines = 800;
 $vote_perday = 3;
@@ -38,11 +40,11 @@ $count_visit = 0;
 
 $fake_info_url = '/book/{aid}/';      
 $fake_chapter_url = '/read/{aid}/{cid}.html';
-$use_orderid = '0';
+$use_orderid = '1';
 $fake_sort_url = '/sort/{sortcode}/{pid}/';      
-$fake_top = '/top.html';        
+$fake_top = '/top/';        
 $fake_recentread = '/history.html';
-$fake_indexlist = '/index/{aid}/{pid}/';  
+$fake_indexlist = '/indexlist/{aid}/{pid}/';  
 $per_indexlist = 50;
 
 //分类设置
@@ -62,6 +64,8 @@ $redisarr = [
     ,'db' => '5'
     ,'pass' => 'caijie123AA'
 ];
+$redis_scope = 'dbpool';
+$redis_pool = '';
 $home_cache_time = 1200;        
 $info_cache_time = 7200;        
 $category_cache_time = 3600;
@@ -84,5 +88,5 @@ $langtail_cache_time = 2592000;
 $fake_langtail_info = '/books/{aid}/';
 $fake_langtail_indexlist = '/indexs/{aid}/{pid}/';
 
-$is_keywords = 0;
+$is_keywords = 1;
 $keywords_num = 5;